fix: move CollectAgentHostDebugLogsAction to electron-browser to fix layer violation
INativeHostService cannot be used from browser/ layer files. Extracted
the action to a new electron-browser/ file and register it from the
existing electron-browser chat contribution.
